pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kodosumi-vibe-template"
version = "0.1.0"
description = "Template for creating CrewAI flows running on a Ray cluster using Kodosumi for serving"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "crewai==0.100.*",
    "ray>=2.6.0",
    "fastapi>=0.95.0",
    "langchain>=0.0.267",
    "langchain-openai>=0.0.2",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "jinja2>=3.0.0",
    "uvicorn>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.0.270",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/kodosumi-vibe-template"
"Bug Tracker" = "https://github.com/yourusername/kodosumi-vibe-template/issues"

[tool.setuptools]
package-dir = {"" = "."}
packages = ["workflows"]

[tool.setuptools.package-data]
"workflows.crewai_flow.templates" = ["*.html"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 88
target-version = "py38"
select = ["E", "F", "B", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py" 